+++ container-engine-lib/docs/changelog.md
@@ -1,5 +1,7 @@
 # TBD
 ### Features
+* Added `KubernetesManager.IsPersistentVolumeClaimBound` to check when a Persistent Volume has been bound to a Persistent Volume Claim
+* Updated `KubernetesManager.CreatePersistentVolumeClaim` now it waits for the Persistent Volume Claim get bound or thrown an error until return
 * Added `CollectMatchingRoles` and `CollectMatchingRoleBindings` in `kubernetes_resource_collectors` package
 * Upped the CircleCI resource class to 'large' since builds are 1m30s and CircleCI showed that we're maxing out the CPU
 * Added a build cache to each build